requirements.txt
asgiref==3.5.2
astroid==2.12.13
attrs==22.1.0
Automat==22.10.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
dill==0.3.6
Django==4.1.4
django-storages==1.13.1
django-stubs==1.13.1
django-stubs-ext==0.7.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
dropbox==11.36.0
filelock==3.8.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
lazy-object-proxy==1.8.0
lxml==4.9.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
parsel==1.7.0
pathspec==0.10.2
Pillow==9.3.0
platformdirs==2.6.0
ply==3.11
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
queuelib==1.6.2
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
scrapy-djangoitem==1.1.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
tldextract==3.4.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
w3lib==2.1.1
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.2
